backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
openai
google-search-results
python-dotenv
httpx
aiosqlite
